+++ apps/admin/src/app/(frontend)/[slug]/page.tsx
@@ -6,8 +6,8 @@ import { cache } from 'react';
 import { RenderBlocks } from '@/lib/blocks/RenderBlocks';
 import { RevealUIRedirects } from '@/lib/components/RevealUIRedirects';
 import { RenderHero } from '@/lib/heros/RenderHero';
-import { generateMeta } from '@/lib/utilities/generateMeta';
-import { getRevealUIInstance } from '@/lib/utilities/revealui-singleton';
+import { generateMeta } from '@/lib/cms/generateMeta';
+import { getRevealUIInstance } from '@/lib/utils/revealui-singleton';
 
 // Force dynamic rendering to prevent build-time RevealUI admin initialization
 export const dynamic = 'force-dynamic';
