+++ apps/admin/src/app/api/gdpr/export/route.ts
@@ -3,8 +3,8 @@ export const runtime = 'nodejs';
 import { getSession } from '@revealui/auth/server';
 import { type NextRequest, NextResponse } from 'next/server';
 import { withRateLimit } from '@/lib/middleware/rate-limit';
-import { writeGDPRAuditEntry } from '@/lib/utilities/gdpr-audit';
-import { getRevealUIInstance } from '@/lib/utilities/revealui-singleton';
+import { writeGDPRAuditEntry } from '@/lib/utils/gdpr-audit';
+import { getRevealUIInstance } from '@/lib/utils/revealui-singleton';
 import { createApplicationErrorResponse, createErrorResponse } from '@/lib/utils/error-response';
 import { extractRequestContext } from '@/lib/utils/request-context';
