+++ apps/admin/src/app/api/memory/context/[sessionId]/[agentId]/route.ts
@@ -13,7 +13,7 @@ import { logger } from '@revealui/utils/logger';
 import { eq } from 'drizzle-orm';
 import { type NextRequest, NextResponse } from 'next/server';
 import { checkAIMemoryFeatureGate } from '@/lib/middleware/ai-feature-gate';
-import { getNodeIdFromSession } from '@/lib/utilities/nodeId';
+import { getNodeIdFromSession } from '@/lib/utils/nodeId';
 import { createErrorResponse, createValidationErrorResponse } from '@/lib/utils/error-response';
 import { extractRequestContext } from '@/lib/utils/request-context';
