+++ apps/admin/src/app/api/memory/episodic/[userId]/[memoryId]/route.ts
@@ -12,7 +12,7 @@ import { getClient } from '@revealui/db/client';
 import { logger } from '@revealui/utils/logger';
 import { type NextRequest, NextResponse } from 'next/server';
 import { checkAIMemoryFeatureGate } from '@/lib/middleware/ai-feature-gate';
-import { getNodeIdFromUser } from '@/lib/utilities/nodeId';
+import { getNodeIdFromUser } from '@/lib/utils/nodeId';
 import {
   createApplicationErrorResponse,
   createErrorResponse,
